Started work on the API
I have created a htaccess file to route all of the requests to the api that do not have a file to the api.php file. From here all of the processing for the api call takes place, directing the call to the appropratie place.

Started work on a REST API for handling the requests.

Updated the README.md to reflect the decision made at the last meeting.