About Section, started SQL
I have added an about section to the app, this features information about the app and how created it us. Needs to be updated with relevent information.

I have started work on the log in portion of the app. currently this is just a simple form and nothing else.

I have started work on the sql required for the DB, this is just the user stuff so I can carry on with current work.