User Creation/login
Updated the README.md to reflect what tools and frameworks are used.

Added in a sql class to handle the connections.

Continued working on the log in page.

Create user now works across the stack. Basic error checking happens in the backend.

Log in now works accross the stack. Basic error checking happens in the backend.

Created a method for checking if a user is logged in.

Updated the DB to include a expiry time on the auth tokens