Bunch of login and user chnages
Fixed bug in UserAPI where was checking if the auth token had expired and returning true otherwise false.

Added a validAuth check into the UserAPI, this is so we can check if a user is still logged in.

Now check if the user is logged in when the app opens.

Created userClass.js to represent a user. Taking advantage of the fact that in js everything is a map to reduce the number of getters and setters. hence the class is quite small. But you need to know what info it is you are asking for.

Now switch from the log in page to the user page on user log in. Also swap the menu and header to reflect the user has logged in.

Updated the static go here pages so they are in line with the rest of the styling.