Templates for places, events, guides
Added in templates for the basic dsiplay of places, events and guides. This is created using handlebard.js to handle the template managing, this can be extended so that we can fill them with the info from the api. I have created temp methods to handle the display of the templates atm, this should be replaced when the functionality is added.